manch-backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "manch-backend"
version = "0.1.0"
description = "Manch orchestration backend — multi-agent AI engineering platform with sandboxed execution"
requires-python = ">=3.11"
license = {text = "MIT"}
readme = "../README.md"
authors = [
  {name = "bigbansal"},
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Framework :: FastAPI",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "orjson>=3.10.0",
  "sqlalchemy>=2.0.32",
  "psycopg[binary]>=3.2.1",
  "opensandbox>=0.1.4",
  "bcrypt>=4.0.0",
  "python-jose[cryptography]>=3.3.0",
  "alembic>=1.13.0",
  "python-multipart>=0.0.9",
]

[project.urls]
Homepage = "https://github.com/bigbansal/manch"
Repository = "https://github.com/bigbansal/manch"
Changelog = "https://github.com/bigbansal/manch/blob/main/CHANGELOG.md"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
